backend/requirements.txt
Django
mysql-connector-python
mysqlclient
djangorestframework
requests
pyrebase4

coverage
